Use unittest for tests
