+++ poc/Makefile
@@ -1,12 +1,2 @@
 test:
-	sage -python common.py
-	sage -python field.py
-	sage -python xof.py
-	sage -python flp.py
-	sage -python flp_generic.py
-	sage -python idpf.py
-	sage -python idpf_poplar.py
-	sage -python daf.py
-	sage -python vdaf.py
-	sage -python vdaf_prio3.py
-	sage -python vdaf_poplar1.py
+	sage -python -m unittest
