+++ poc/idpf_poplar.py
@@ -3,10 +3,10 @@
 import itertools
 
 import field
-from common import (ERR_DECODE, ERR_INPUT, TEST_VECTOR, Bytes, Unsigned, Vec,
-                    format_dst, vec_add, vec_neg, vec_sub, xor)
+from common import (ERR_DECODE, ERR_INPUT, Bytes, Unsigned, Vec, format_dst,
+                    vec_add, vec_neg, vec_sub, xor)
 from field import Field2
-from idpf import Idpf, gen_test_vec, test_idpf, test_idpf_exhaustive
+from idpf import Idpf
 from xof import XofFixedKeyAes128
 
 
@@ -265,26 +265,3 @@ def unpack_bits(packed_bits: Bytes, length: Unsigned) -> Vec[Field2]:
     if (length + 7) // 8 != len(packed_bits) or leftover_bits != 0:
         raise ERR_DECODE
     return bits
-
-
-if __name__ == '__main__':
-    cls = IdpfPoplar \
-        .with_value_len(2)
-    if TEST_VECTOR:
-        gen_test_vec(cls.with_bits(10), 0, 0)
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 15, (0b1111000011110000,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 14, (0b111100001111000,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 13, (0b11110000111100,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 12, (0b1111000011110,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 11, (0b111100001111,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 10, (0b11110000111,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 5, (0b111100,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 4, (0b11110,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 3, (0b1111,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 2, (0b111,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 1, (0b11,))
-    test_idpf(cls.with_bits(16), 0b1111000011110000, 0, (0b1,))
-    test_idpf(cls.with_bits(1000), 0, 999, (0,))
-    test_idpf_exhaustive(cls.with_bits(1), 0)
-    test_idpf_exhaustive(cls.with_bits(1), 1)
-    test_idpf_exhaustive(cls.with_bits(8), 91)
