+++ poc/vdaf_prio3.py
@@ -5,10 +5,10 @@ from typing import Optional, Union
 import flp
 import flp_generic
 import xof
-from common import (ERR_INPUT, ERR_VERIFY, TEST_VECTOR, Unsigned, byte, concat,
-                    front, vec_add, vec_sub, zeros)
+from common import (ERR_INPUT, ERR_VERIFY, Unsigned, byte, concat, front,
+                    vec_add, vec_sub, zeros)
 from field import FftField, Field64, Field128
-from vdaf import Vdaf, test_vdaf
+from vdaf import Vdaf
 
 USAGE_MEAS_SHARE = 1
 USAGE_PROOF_SHARE = 2
@@ -571,197 +571,3 @@ class Prio3MultiHotHistogram(Prio3):
                 length, max_count, chunk_length
             ))
         return Prio3MultiHotHistogramWithParams
-
-
-##
-# TESTS
-#
-
-class TestPrio3Average(Prio3):
-    """
-    A Prio3 instantiation to test use of num_measurements in the Valid
-    class's decode() method.
-    """
-
-    Xof = xof.XofTurboShake128
-    # NOTE 0xFFFFFFFF is reserved for testing. If we decide to standardize this
-    # Prio3 variant, then we'll need to pick a real codepoint for it.
-    ID = 0xFFFFFFFF
-    VERIFY_KEY_SIZE = xof.XofTurboShake128.SEED_SIZE
-
-    @classmethod
-    def with_bits(cls, bits: Unsigned):
-        class TestPrio3AverageWithBits(TestPrio3Average):
-            Flp = flp_generic.FlpGeneric(flp_generic.TestAverage(bits))
-        return TestPrio3AverageWithBits
-
-
-def _test_prio3sumvec(num_proofs: Unsigned, field: FftField):
-    valid_cls = flp_generic.SumVec.with_field(field)
-    assert Prio3SumVecWithMultiproof.is_recommended(
-        valid_cls, num_proofs, field)
-
-    cls = Prio3SumVecWithMultiproof \
-        .with_params(10, 8, 9, num_proofs, field) \
-        .with_shares(2)
-
-    assert cls.ID == 0xFFFFFFFF
-    assert cls.PROOFS == num_proofs
-
-    test_vdaf(
-        cls,
-        None,
-        [[1, 61, 86, 61, 23, 0, 255, 3, 2, 1]],
-        [1, 61, 86, 61, 23, 0, 255, 3, 2, 1]
-    )
-    test_vdaf(
-        cls,
-        None,
-        [
-            list(range(10)),
-            [1] * 10,
-            [255] * 10
-        ],
-        list(range(256, 266)),
-        print_test_vec=False,
-    )
-    cls = Prio3SumVec.with_params(3, 16, 7).with_shares(3)
-    test_vdaf(
-        cls,
-        None,
-        [
-            [10000, 32000, 9],
-            [19342, 19615, 3061],
-            [15986, 24671, 23910]
-        ],
-        [45328, 76286, 26980],
-        print_test_vec=False,
-        test_vec_instance=1,
-    )
-
-
-def test_prio3sumvec_with_multiproof():
-    for n in range(2, 5):
-        _test_prio3sumvec(num_proofs=n, field=Field64)
-
-
-if __name__ == '__main__':
-    num_shares = 2  # Must be in range `[2, 256)`
-
-    cls = Prio3 \
-        .with_xof(xof.XofTurboShake128) \
-        .with_flp(flp.FlpTestField128()) \
-        .with_shares(num_shares)
-    cls.ID = 0xFFFFFFFF
-    test_vdaf(cls, None, [1, 2, 3, 4, 4], 14)
-
-    # If JOINT_RAND_LEN == 0, then Fiat-Shamir isn't needed and we can skip
-    # generating the joint randomness.
-    cls = Prio3 \
-        .with_xof(xof.XofTurboShake128) \
-        .with_flp(flp.FlpTestField128.with_joint_rand_len(0)) \
-        .with_shares(num_shares)
-    cls.ID = 0xFFFFFFFF
-    test_vdaf(cls, None, [1, 2, 3, 4, 4], 14)
-
-    cls = Prio3Count.with_shares(num_shares)
-    assert cls.ID == 0x00000000
-    test_vdaf(cls, None, [0, 1, 1, 0, 1], 3)
-    test_vdaf(cls, None, [1], 1, print_test_vec=TEST_VECTOR)
-    cls = Prio3Count.with_shares(3)
-    test_vdaf(cls, None, [1], 1, print_test_vec=TEST_VECTOR,
-              test_vec_instance=1)
-
-    cls = Prio3Sum.with_bits(8).with_shares(num_shares)
-    assert cls.ID == 0x00000001
-    test_vdaf(cls, None, [0, 147, 1, 0, 11, 0], 159)
-    test_vdaf(cls, None, [100], 100, print_test_vec=TEST_VECTOR)
-    cls = Prio3Sum.with_bits(8).with_shares(3)
-    test_vdaf(cls, None, [100], 100, print_test_vec=TEST_VECTOR,
-              test_vec_instance=1)
-
-    cls = Prio3SumVec.with_params(10, 8, 9).with_shares(2)
-    assert cls.ID == 0x00000002
-    test_vdaf(
-        cls,
-        None,
-        [[1, 61, 86, 61, 23, 0, 255, 3, 2, 1]],
-        [1, 61, 86, 61, 23, 0, 255, 3, 2, 1]
-    )
-    test_vdaf(
-        cls,
-        None,
-        [
-            list(range(10)),
-            [1] * 10,
-            [255] * 10
-        ],
-        list(range(256, 266)),
-        print_test_vec=TEST_VECTOR,
-    )
-    cls = Prio3SumVec.with_params(3, 16, 7).with_shares(3)
-    test_vdaf(
-        cls,
-        None,
-        [
-            [10000, 32000, 9],
-            [19342, 19615, 3061],
-            [15986, 24671, 23910]
-        ],
-        [45328, 76286, 26980],
-        print_test_vec=TEST_VECTOR,
-        test_vec_instance=1,
-    )
-
-    cls = Prio3Histogram \
-        .with_params(4, 2) \
-        .with_shares(num_shares)
-    assert cls.ID == 0x00000003
-    test_vdaf(cls, None, [0], [1, 0, 0, 0])
-    test_vdaf(cls, None, [1], [0, 1, 0, 0])
-    test_vdaf(cls, None, [2], [0, 0, 1, 0])
-    test_vdaf(cls, None, [3], [0, 0, 0, 1])
-    test_vdaf(cls, None, [0, 0, 1, 1, 2, 2, 3, 3], [2, 2, 2, 2])
-    test_vdaf(cls, None, [2], [0, 0, 1, 0], print_test_vec=TEST_VECTOR)
-    cls = Prio3Histogram.with_params(11, 3).with_shares(3)
-    test_vdaf(
-        cls,
-        None,
-        [2],
-        [0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0],
-        print_test_vec=TEST_VECTOR,
-        test_vec_instance=1,
-    )
-
-    # Prio3MultiHotHistogram with length = 4, max_count = 2, chunk_length = 2.
-    cls = Prio3MultiHotHistogram \
-        .with_params(4, 2, 2) \
-        .with_shares(num_shares)
-    assert cls.ID == 0xFFFFFFFF
-    test_vdaf(cls, None, [[0, 0, 0, 0]], [0, 0, 0, 0])
-    test_vdaf(cls, None, [[0, 1, 0, 0]], [0, 1, 0, 0])
-    test_vdaf(cls, None, [[0, 1, 1, 0]], [0, 1, 1, 0])
-    test_vdaf(cls, None, [[0, 1, 1, 0], [0, 1, 0, 1]], [0, 2, 1, 1])
-    test_vdaf(
-        cls, None, [[0, 1, 1, 0]], [0, 1, 1, 0], print_test_vec=False
-    )
-    # Prio3MultiHotHistogram with length = 11, max_count = 5, chunk_length = 3.
-    cls = Prio3MultiHotHistogram.with_params(11, 5, 3).with_shares(3)
-    test_vdaf(
-        cls,
-        None,
-        [[1] * 5 + [0] * 6],
-        [1] * 5 + [0] * 6,
-        print_test_vec=False,
-        test_vec_instance=1,
-    )
-
-    cls = TestPrio3Average.with_bits(3).with_shares(num_shares)
-    test_vdaf(cls, None, [1, 5, 1, 1, 4, 1, 3, 2], 2)
-
-    # Test `is_valid` returns True on empty previous_agg_params, and False
-    # otherwise.
-    assert cls.is_valid(None, set([]))
-    assert not cls.is_valid(None, set([None]))
-
-    test_prio3sumvec_with_multiproof()
